테스트 서비스 URL 공통 컨텍스트로 분리
