+++ src/test/java/com/camping/tests/steps/KioskSteps.java
@@ -4,39 +4,27 @@ import io.cucumber.java.ko.그러면;
 import io.cucumber.java.ko.만약;
 import io.restassured.response.Response;
 
-import java.util.Map;
-
 import static io.restassured.RestAssured.given;
 import static org.hamcrest.Matchers.*;
 
 public class KioskSteps {
-    private static final Map<String, String> SERVICE_URLS = Map.of(
-            "kiosk", System.getProperty("kiosk.base.url",
-                    System.getenv().getOrDefault("KIOSK_BASE_URL", "http://localhost:18080")),
-            "admin", System.getProperty("admin.base.url",
-                    System.getenv().getOrDefault("ADMIN_BASE_URL", "http://localhost:18081")),
-            "reservation", System.getProperty("reservation.base.url",
-                    System.getenv().getOrDefault("RESERVATION_BASE_URL", "http://localhost:18082"))
-    );
-
-    private final TestContext context;
+    private final CommonContext context;
     private Response response;
 
-    public KioskSteps(TestContext context) {
+    public KioskSteps(CommonContext context) {
         this.context = context;
     }
 
     @만약("상품 목록을 조회한다")
-    public void getProducts() {
-        response =
-                given()
-                        .header("Authorization", "Bearer " + context.authToken())
-                        .when()
-                        .get(SERVICE_URLS.get("kiosk") + "/api/products");
+    public void 상품_목록을_조회한다() {
+        response = given()
+//                .header("Authorization", "Bearer " + context.authToken())
+                .when()
+                .get(context.serviceUrl("kiosk") + "/api/products");
     }
 
     @그러면("상품 목록이 정상적으로 조회된다")
-    public void verifyProducts() {
+    public void 상품_목록이_정상적으로_조회된다() {
         response.then()
                 .statusCode(200)
                 .body("$.size()", greaterThanOrEqualTo(1))
